+++ primitives/order/Cargo.toml
@@ -9,8 +9,8 @@ edition = "2021"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-async-trait = { workspace = true }
-log = { workspace = true }
+async-trait = { workspace = true, optional = true }
+log = { workspace = true, default-features = false }
 hex-literal = { workspace = true }
 codec = { workspace = true, default-features = false, features = ["derive"] }
 scale-info = { workspace = true, default-features = false, features = ["derive"] }
@@ -27,7 +27,7 @@ sp-state-machine = { workspace = true, default-features = false }
 
 # Cumulus
 cumulus-primitives-core = { workspace = true, default-features = false }
-cumulus-relay-chain-interface = { workspace = true, default-features = false }
+cumulus-relay-chain-interface = { workspace = true, default-features = false, optional = true }
 
 [dev-dependencies]
 serde = { workspace = true }
@@ -36,6 +36,8 @@ smallvec = { workspace = true }
 [features]
 default = ["std"]
 std = [
+    "async-trait",
+    "log/std",
     "codec/std",
     "scale-info/std",
     "frame-support/std",
@@ -47,4 +49,5 @@ std = [
     "sp-trie/std",
     "sp-state-machine/std",
     "cumulus-primitives-core/std",
+    "cumulus-relay-chain-interface"
 ]
